refact : 추천 도메인 분리
profiles -> recomendation
