+++ repo/profiles/urls.py
@@ -18,7 +18,6 @@ urlpatterns = [
     path("follow/", views.FollowListAPIView.as_view(), name="my_follow_list"),
     path("<int:id>/block/", views.BlockListCreateDeleteAPIView.as_view(), name="my_block_list"),
     path("block/", views.BlockListAPIView.as_view(), name="my_block_list"),
-    path("recommend/", views.BudyRecommendAPIView.as_view(), name="budy-recommend"),
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("<int:id>/posts/", views.UserPostListAPIView.as_view(), name="user_posts"),
