+++ repo/profiles/views.py
@@ -1,5 +1,3 @@
-import random
-
 import jwt
 import requests
 from allauth.socialaccount.providers.kakao import views as kakao_view
@@ -8,7 +6,6 @@ from allauth.socialaccount.providers.oauth2.client import OAuth2Client
 from dj_rest_auth.registration.views import SocialLoginView
 from django.conf import settings
 from django.db import transaction
-from django.db.models import Count, Q
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from django.utils.timezone import now
@@ -25,7 +22,6 @@ from repo.common.utils import get_first_photo_url, get_paginated_response_with_c
 from repo.profiles.models import CustomUser, Relationship, UserDetail
 from repo.profiles.schemas import *
 from repo.profiles.serializers import (
-    BudyRecommendSerializer,
     UserBlockListSerializer,
     UserDetailSignupSerializer,
     UserFollowListSerializer,
@@ -489,50 +485,6 @@ class BlockListCreateDeleteAPIView(APIView):
         if not is_deleted:
             return Response({"error": "User is not blocking"}, status=status.HTTP_404_NOT_FOUND)
         return Response({"success": "unblock"}, status=status.HTTP_200_OK)
-
-
-@BudyRecommendSchema.budy_recommend_schema_view
-class BudyRecommendAPIView(APIView):
-    """
-    유저의 커피 즐기는 방식 6개 중 한가지 방식에 해당 하는 유저 리스트 반환
-    Args:
-        request: 클라이언트로부터 받은 요청 객체
-    Returns:
-        users:
-            user: 유저의 커피 생활 방식에 해당 하는 유저 리스트 반환 (10명)
-            follower_cnt: 유저의 팔로워 수
-        category: 커피 생활 방식
-        실패 시: HTTP 404 Not Found
-
-    담당자: hwtar1204
-    """
-
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        user = request.user
-        user_detail = get_object_or_404(UserDetail, user=user)
-        coffee_life_helper = user_detail.get_coffee_life_helper()
-        true_categories = coffee_life_helper.get_true_categories()
-
-        if not true_categories:
-            category = random.choice(UserDetail.COFFEE_LIFE_CHOICES)
-        else:
-            category = random.choice(true_categories)
-
-        recommend_user_list = (
-            CustomUser.objects.select_related("user_detail")
-            .only("user_detail__coffee_life")
-            .filter(user_detail__coffee_life__contains={category: True})
-            .exclude(id=user.id)
-            .annotate(follower_cnt=Count("relationships_to", filter=Q(relationships_to__relationship_type="follow")))
-            .order_by("?")[:10]
-        )
-
-        serializer = BudyRecommendSerializer(recommend_user_list, many=True)
-        response_data = {"users": serializer.data, "category": category}
-
-        return Response(response_data, status=status.HTTP_200_OK)
 
 
 @UserPostListSchema.user_post_list_schema_view
